Modify the clone function of agent to reflect options and cut out to clone.go
